+++ control/src/cache/reconcile/heads.rs
@@ -2,7 +2,6 @@ use super::{calculate, taper};
 use crate::cache::display::AttributeScreen;
 use crate::cache::mapping::AttributeIdentifier;
 use crate::cache::quantization::{quantize, Quantization};
-use crate::cache::AttributesHead;
 use crate::Store;
 
 impl Store {
@@ -11,14 +10,12 @@ impl Store {
             self.reconcile_position(i);
         }
 
-        let ordered_heads = self.heads_ordered_by_position();
-        self.set_screen_for_positions(&ordered_heads);
-        let relative_position_by_index = relative_position_by_index(&ordered_heads);
+        self.set_screen_for_positions();
 
-        for (i, relative_position) in relative_position_by_index.iter().enumerate() {
-            self.reconcile_volume(i, *relative_position);
-            self.reconcile_feedback(i, *relative_position);
-            self.reconcile_pan(i, *relative_position);
+        for i in 0..4 {
+            self.reconcile_volume(i);
+            self.reconcile_feedback(i);
+            self.reconcile_pan(i);
         }
     }
 
@@ -34,7 +31,7 @@ impl Store {
         );
     }
 
-    fn reconcile_volume(&mut self, i: usize, relative_position: usize) {
+    fn reconcile_volume(&mut self, i: usize) {
         let volume_sum = super::sum(
             (self.input.head[i].volume.value() - 0.02) / 0.98,
             self.control_value_for_attribute(AttributeIdentifier::Volume(i)),
@@ -42,7 +39,7 @@ impl Store {
         // The top limit is made to match compressor's treshold.
         self.cache.attributes.head[i].volume =
             super::calculate_from_sum(volume_sum, (0.0, 0.25), Some(taper::log));
-        let screen = AttributeScreen::Volume(i, relative_position, volume_sum);
+        let screen = AttributeScreen::Volume(i, volume_sum);
         if self.input.head[i].volume.active() {
             self.cache.display.force_attribute(screen);
         } else {
@@ -50,14 +47,14 @@ impl Store {
         }
     }
 
-    fn reconcile_feedback(&mut self, i: usize, relative_position: usize) {
+    fn reconcile_feedback(&mut self, i: usize) {
         let feedback_sum = super::sum(
             (self.input.head[i].feedback.value() - 0.02) / 0.98,
             self.control_value_for_attribute(AttributeIdentifier::Feedback(i)),
         );
         self.cache.attributes.head[i].feedback =
             super::calculate_from_sum(feedback_sum, (0.0, 1.2), None);
-        let screen = AttributeScreen::Feedback(i, relative_position, feedback_sum);
+        let screen = AttributeScreen::Feedback(i, feedback_sum);
         if self.input.head[i].feedback.active() {
             self.cache.display.force_attribute(screen);
         } else {
@@ -65,13 +62,13 @@ impl Store {
         }
     }
 
-    fn reconcile_pan(&mut self, i: usize, relative_position: usize) {
+    fn reconcile_pan(&mut self, i: usize) {
         let pan_sum = super::sum(
             self.input.head[i].pan.value(),
             self.control_value_for_attribute(AttributeIdentifier::Pan(i)),
         );
         self.cache.attributes.head[i].pan = super::calculate_from_sum(pan_sum, (0.0, 1.0), None);
-        let screen = AttributeScreen::Pan(i, relative_position, pan_sum);
+        let screen = AttributeScreen::Pan(i, pan_sum);
         if self.input.head[i].pan.active() {
             self.cache.display.force_attribute(screen);
         } else {
@@ -79,25 +76,8 @@ impl Store {
         }
     }
 
-    fn heads_ordered_by_position(&self) -> [(usize, AttributesHead); 4] {
-        let mut ordered_heads = [
-            (0, self.cache.attributes.head[0]),
-            (1, self.cache.attributes.head[1]),
-            (2, self.cache.attributes.head[2]),
-            (3, self.cache.attributes.head[3]),
-        ];
-        for i in 0..ordered_heads.len() {
-            for j in 0..ordered_heads.len() - 1 - i {
-                if ordered_heads[j].1.position > ordered_heads[j + 1].1.position {
-                    ordered_heads.swap(j, j + 1);
-                }
-            }
-        }
-        ordered_heads
-    }
-
-    fn set_screen_for_positions(&mut self, ordered_heads: &[(usize, AttributesHead); 4]) {
-        let screen_for_positions = screen_for_positions(ordered_heads);
+    fn set_screen_for_positions(&mut self) {
+        let screen_for_positions = self.screen_for_positions();
         let touched_position = self.input.head.iter().any(|h| h.position.active());
         if touched_position {
             self.cache.display.force_attribute(screen_for_positions);
@@ -108,29 +88,22 @@ impl Store {
             .display
             .set_fallback_attribute(screen_for_positions);
     }
-}
 
-fn screen_for_positions(ordered_heads: &[(usize, AttributesHead); 4]) -> AttributeScreen {
-    AttributeScreen::Positions((
-        [
-            ordered_heads[0].1.volume > 0.00,
-            ordered_heads[1].1.volume > 0.00,
-            ordered_heads[2].1.volume > 0.00,
-            ordered_heads[3].1.volume > 0.00,
-        ],
-        [
-            ordered_heads[0].1.feedback > 0.00,
-            ordered_heads[1].1.feedback > 0.00,
-            ordered_heads[2].1.feedback > 0.00,
-            ordered_heads[3].1.feedback > 0.00,
-        ],
-    ))
-}
-
-fn relative_position_by_index(ordered_heads: &[(usize, AttributesHead); 4]) -> [usize; 4] {
-    let mut relative_position_by_index = [0; 4];
-    for (relative_position, (i, _)) in ordered_heads.iter().enumerate() {
-        relative_position_by_index[*i] = relative_position;
+    fn screen_for_positions(&self) -> AttributeScreen {
+        // TODO: Handle hysteresis for position
+        AttributeScreen::Positions((
+            [
+                self.cache.attributes.head[0].volume > 0.00,
+                self.cache.attributes.head[1].volume > 0.00,
+                self.cache.attributes.head[2].volume > 0.00,
+                self.cache.attributes.head[3].volume > 0.00,
+            ],
+            [
+                self.cache.attributes.head[0].feedback > 0.00,
+                self.cache.attributes.head[1].feedback > 0.00,
+                self.cache.attributes.head[2].feedback > 0.00,
+                self.cache.attributes.head[3].feedback > 0.00,
+            ],
+        ))
     }
-    relative_position_by_index
 }